+++ server/api/shops/[id].delete.ts
@@ -7,12 +7,26 @@ export default defineEventHandler(async (event) => {
       setResponseStatus(event, 400);
       return { error: "Bad Request" };
     }
-    const shop = await prisma.shop.delete({
+    const userEmail = event.req.headers["user-email"] as string;
+    if (!userEmail) {
+      setResponseStatus(event, 400);
+      return { error: "User email is required in header as 'user-email'" };
+    }
+    const shop = await prisma.shop.findUnique({
+      where: {
+        userEmail,
+      },
+    });
+    if (!shop || shop.userEmail !== userEmail) {
+      setResponseStatus(event, 404);
+      return { error: "You are not the shop owner" };
+    }
+    const deletedShop = await prisma.shop.delete({
       where: {
         id,
       },
     });
-    return shop;
+    return deletedShop;
   } catch (error) {
     console.error("Error deleting shop:", error);
     setResponseStatus(event, 500);
